+++ Examples/Resources/WaitForADDomain/7-WaitForADDomain_WaitForDomainControllerIgnoringAuthenticationErrors_Config.ps1
@@ -39,10 +39,10 @@ Configuration WaitForADDomain_WaitForDomainControllerIgnoringAuthenticationError
     {
         WaitForADDomain 'contoso.com'
         {
-            DomainName                 = 'contoso.com'
-            IgnoreAuthenticationErrors = $true
+            DomainName             = 'contoso.com'
+            WaitForValidCredentials = $true
 
-            PsDscRunAsCredential       = $Credential
+            PsDscRunAsCredential   = $Credential
         }
     }
 }